Working on registering the whole stack, created a stub of the upcoming function, corrected a few mistakes in the resizing and installed MIJ

+++ For3D/resize_tif.m
@@ -23,7 +23,7 @@ if nargin<3, Amax = 0; Bmax = 0; end
 
 dir_out = 'resized'; % '../resized';
 
-if (~iscell(file))
+if (~iscell(files))
   [filepath, filename, fileext] = fileparts(files);
   ls = dir(files);
 
@@ -46,7 +46,7 @@ if (N == 0), disp('nada??'), return, end
 
 if (Amax == 0)
     A = zeros(N, 1); B = A;
-    
+
     for i = 1:N % loop over images to find max size
         filename = files{i};
         im = imread(filename);
@@ -64,35 +64,41 @@ new_names = files;
 for i = 1:N % loop over images to resize images
     filename = files{i};
     im = imread(filename);
-    
+
     [a, b, c] = size(im);
 %     if b>b_max, im = imresize(im, 1/2); [a, b, c] = size(im); fprintf('image too large, %i pxl width, resized\r', b), end
-    if (a < Amax || b < Bmax)
-      [filepath, fname, fileext] = fileparts(filename);
-      out_path = fullfile(filepath, dir_out);
+    [filepath, fname, fileext] = fileparts(filename);
+    out_path = fullfile(filepath, dir_out);
 
-      if ~isdir(out_path)
-        mkdir(out_path);
-      end
-      new_name = fullfile(outpath, [fname fileext]);
-      
-      border_rows = squeeze([im(1,:,:) im(end,:,:)]);
-      border_cols = squeeze([im(:,1,:); im(:,end,:)]);
-      background_val = mean([border_rows; border_cols]);
-      
-      fprintf('%s\n', filename)
-      a0 = round(Amax/2-a/2);
-      b0 = round(Bmax/2-b/2);
-      
-      im2 = zeros(Amax, Bmax, c, class(im));
-      for j = 1:c
-          im2(:,:,j) = im2(:,:,j) + background_val(j);
-      end
-      im2(a0+1:a0+a, b0+1:b0+b, :) = im;
-      imwrite(im2, new_name, 'Compression', 'none')
-
-      new_names{i} = new_name;
+    if ~isdir(out_path)
+      mkdir(out_path);
     end
+    new_name = fullfile(out_path, [fname fileext]);
+
+    fprintf('%s\n', new_name);
+
+    if (~exist(new_name, 'file'))
+      if (a < Amax || b < Bmax)
+
+        border_rows = squeeze([im(1,:,:) im(end,:,:)]);
+        border_cols = squeeze([im(:,1,:); im(:,end,:)]);
+        background_val = median([border_rows; border_cols]);
+
+        a0 = round(Amax/2-a/2);
+        b0 = round(Bmax/2-b/2);
+
+        im2 = zeros(Amax, Bmax, c, class(im));
+        for j = 1:c
+            im2(:,:,j) = im2(:,:,j) + background_val(j);
+        end
+        im2(a0+1:a0+a, b0+1:b0+b, :) = im;
+        imwrite(im2, new_name, 'Compression', 'none')
+
+      else
+        copyfile(filename, new_name);
+      end
+    end
+    new_names{i} = new_name;
 end
 
 %%%
